pyproject.toml
[tool.poetry]
name = "nonprofit-engagement-model"
version = "0.1.0"
description = "A Python analytics project implementing BG/NBD model for predicting supporter engagement in non-profit organizations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.0"
numpy = "^1.24.0"
pymc = "^5.0.0"
arviz = "^0.18.0"
sqlalchemy = "^2.0.0"
pyodbc = "^4.0.0"
python-dotenv = "^1.0.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
scikit-learn = "^1.3.0"
jupyter = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88